# backend/docs/openapi.yml
openapi: 3.1.0
info:
    title: P2P Streaming service - API 
    version: 1.0.0
    description:
        It is streaming service based on bittorrent. We take the data from the torrent.
        
        
        It allows users to watch movies available on the service, create playlists and do a lot more. Administrators can add new movies and edit old ones.
servers: 
  - url: http://localhost/api/v1
...
paths:
    /movies:
        get: 
          tags: 
            - Movies
          summary: Getting the list of movies
          description: Getting the list of movie with pagination
          operationId: getMovies
          parameters: 
            - name: limit
              in: query
              description: Limit of films
              required: false
              schema:
                type: integer
                format: int64 
            - name: offset
              in: query
              description: Offset of films
              required: false
              schema:
                type: integer
                format: int64 
          responses: 
            '200':
              description: Successful getting of the movies
              content: 
                application/json: 
                  schema: 
                    type: array
                    items: 
                      $ref: '#/components/schemas/Movie'
            '400':
              description: Invalid Limit or Offset
              content:
                application/json: 
                  schema: 
                    $ref: '#/components/schemas/Error' 
...
components:
  schemas: 
    Movie:
      type: object
      properties:
        id:
          type: integer
          format: int64
          examples: [1]
        name:
          type: string
          examples: ["Иван Васильевич меняет профессию"]
...
    Error:
      type: object
      properties:
        error:
          type: string
          examples: ["Incorrect data."]
  securitySchemes: 
    ApiKey:
      description: Jwt api key
      type: apiKey
      name: api-key
      in: header